HPCC-12127 Use 64-bit integer for HTTP Content-Length header

+++ esp/platform/espcontext.cpp
@@ -418,7 +418,7 @@ public:
         m_traceValues.append(str.str());
     }
 
-    virtual void addTraceSummaryValue(const char *name, int value)
+    virtual void addTraceSummaryValue(const char *name, __int64 value)
     {
         StringBuffer str;
         if (name && *name)
